Move test to separate test class & add CHANGES entry
